pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phronesis-index"
version = "1.0.0"
description = "Spectral Sheaf Heuristics for Consistency Detection in Multi-Agent Systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Sepehr Bayat", email = "sepehrbayat@hooshex.com"},
]
keywords = ["sheaf", "topology", "multi-agent", "consistency", "spectral"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "networkx>=2.6.0",
    "matplotlib>=3.4.0",
]

[project.optional-dependencies]
experiments = [
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.62.0",
    "seaborn>=0.11.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["phronesis*"]
